project1

+++ backend/warehouses/src/main/java/com/skillstorm/warehouses/services/WarehouseService.java
@@ -8,6 +8,7 @@ import com.skillstorm.warehouses.repositories.WarehouseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
+import java.util.Optional;
 
 @org.springframework.stereotype.Service
 public class WarehouseService {
@@ -17,9 +18,15 @@ public class WarehouseService {
 
     Message m = new Message();
 
-    public Warehouse newWarehouse(Warehouse warehouse) {
+    public Message newWarehouse(Warehouse warehouse) {
+            try{  warehouseRepository.save(warehouse);
+                m.setInfo("saved");
+        return m;
+                }catch (Exception e ){
+                m.setInfo("invalid input");
 
-        return warehouseRepository.save(warehouse);
+                return m;
+            }
     }
     public List<Warehouse> findAllwarehoses() {
         return warehouseRepository.findAll();
@@ -27,13 +34,18 @@ public class WarehouseService {
 
     public Message deleteById(int id) {
         try{
-
+          Optional<Warehouse> w  = warehouseRepository.findById(id);
+            System.out.println("wae" + w);
+            if (w.get().getElectronic().size() != 0){
+                m.setInfo("warehouse: " + id + " has items please move all item to another warehouse" );
+                return m;
+            }
             m.setInfo(" warehouse: " + id +" has been deleted");
             warehouseRepository.deleteById(id);
             return m;
         }
         catch (Exception e) {
-            m.setInfo("warehouse: " + id + " not found" );
+            m.setInfo("warehouse: " + id + " doesn't exist" );
 
         }
